feat: implement adoption request feature with endpoints for creating, retrieving, and updating requests
- Added new endpoints for adoption requests in the API
- Created AdoptionRequestController to handle requests
- Implemented AdoptionRequestService for business logic
- Introduced AdoptionRequest entity and repository for persistence
- Added DTOs for request and response handling
- Updated PetService to calculate age with decimal precision
- Enhanced README and guides with adoption request details
- Added tests for adoption request service functionality

+++ src/main/java/pawtner_core/pawtner_care_api/pet/dto/PetResponse.java
@@ -13,7 +13,7 @@ public record PetResponse(
     BigDecimal weight,
     BigDecimal height,
     LocalDate birthDate,
-    Integer age,
+    BigDecimal age,
     LocalDate adoptionDate,
     PetAdopterResponse adoptedBy,
     LocalDate rescuedDate,
